docker-compose, controllers, services, mappings, stuff)

+++ src/main/java/ua/edu/ukma/event_management_system/user/internal/UserDto.java
@@ -9,7 +9,7 @@ import java.time.LocalDate;
 @NoArgsConstructor
 @Data
 public class UserDto {
-    private long id;
+    private Long id;
     private UserRole userRole;
     @NotBlank(message = "Username is required")
     private String username;
@@ -25,7 +25,7 @@ public class UserDto {
     private String phoneNumber;
     private LocalDate dateOfBirth;
 
-    public UserDto(int id, UserRole userRole, String username, String firstName, String lastName, String email, String phoneNumber, LocalDate dateOfBirth) {
+    public UserDto(Long id, UserRole userRole, String username, String firstName, String lastName, String email, String phoneNumber, LocalDate dateOfBirth) {
         this.id = id;
         this.userRole = userRole;
         this.username = username;
@@ -36,7 +36,7 @@ public class UserDto {
         this.dateOfBirth = dateOfBirth;
     }
 
-    public UserDto(int id, UserRole userRole, String username, String firstName, String lastName, String email, String password, String phoneNumber, LocalDate dateOfBirth) {
+    public UserDto(Long id, UserRole userRole, String username, String firstName, String lastName, String email, String password, String phoneNumber, LocalDate dateOfBirth) {
         this.id = id;
         this.userRole = userRole;
         this.username = username;
